tests: fix tests for service classes
