+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/data/repo/Repo_MySql_Tests.kt
@@ -14,6 +14,9 @@ package test.data.repo
 import kotlinx.coroutines.runBlocking
 import kiit.common.data.*
 import kiit.entities.*
+import org.junit.Assert
+import org.junit.Ignore
+import org.junit.Test
 import test.entities.EntitySetup
 import test.setup.*
 
@@ -27,7 +30,7 @@ Sql scripts located :
 open class Repo_MySql_Tests : Repo_Common_Tests() {
 
     /**
-     * This provices the implementation for the repo ( using mysql as vendor )
+     * This provides the implementation for the repo ( using mysql as vendor )
      */
     override fun process(op: (EntityRepo<Long, User5>) -> Unit) {
         val db = EntitySetup.db(Vendor.MySql)
