+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/entities/Entity_Service_Types_Tests.kt
@@ -18,21 +18,19 @@ import org.junit.Assert
 import org.junit.Test
 import org.threeten.bp.*
 import kiit.common.DateTimes
-import kiit.common.data.Connections
 import kiit.common.data.Vendor
 import kiit.common.ids.UPIDs
 import kiit.data.core.LongId
 import kiit.db.Db
 import kiit.entities.EntityService
 import kiit.query.Op
-import org.junit.Ignore
 import test.TestApp
 import test.setup.Address
 import test.setup.StatusEnum
 import test.setup.TestSupport
 import java.util.*
 
-class Data_04_Entity_Service_Types : TestSupport {
+class Entity_Service_Types_Tests : TestSupport {
 
     val sampleUUID1 = "67bdb72a-1d74-11e8-b467-0ed5f89f7181"
      val sampleUUID2 = "67bdb72a-1d74-11e8-b467-0ed5f89f7182"
